add validation in product default image delete

+++ src/main/java/com/nice/service/impl/ProductServiceImpl.java
@@ -83,7 +83,7 @@ import com.nice.util.ExportCSV;
 
 /**
  * @author : Kody Technolab PVT. LTD.
- * @date : 29-Jun-2020
+ * @date   : 29-Jun-2020
  */
 @Service(value = "productService")
 @Transactional(rollbackFor = Throwable.class)
@@ -386,7 +386,7 @@ public class ProductServiceImpl implements ProductService {
 	/**
 	 * validation for add or update product
 	 *
-	 * @param productRequestDTO
+	 * @param  productRequestDTO
 	 * @throws NotFoundException
 	 * @throws ValidationException
 	 */
@@ -495,13 +495,12 @@ public class ProductServiceImpl implements ProductService {
 	}
 
 	/**
-	 * listForAdmin==null means get product detail for admin listForAdmin==true means get product list for admin convert
-	 * entity to response dto
+	 * listForAdmin==null means get product detail for admin listForAdmin==true means get product list for admin convert entity to response dto
 	 *
-	 * @param product
-	 * @param listForAdmin
-	 * @param productParamRequestDTO
-	 * @param pincodeId
+	 * @param  product
+	 * @param  listForAdmin
+	 * @param  productParamRequestDTO
+	 * @param  pincodeId
 	 * @return
 	 * @throws NotFoundException
 	 * @throws ValidationException
@@ -602,8 +601,8 @@ public class ProductServiceImpl implements ProductService {
 		productResponseDTO.setProductAvailable(makeProductVisible);
 
 		/**
-		 * Here flag is set to determine whether to make product visible to customer or not, if the setProductAvailable flag is
-		 * true then product should be made visible else that product should be shown as out of stock
+		 * Here flag is set to determine whether to make product visible to customer or not, if the setProductAvailable flag is true then product should be made
+		 * visible else that product should be shown as out of stock
 		 */
 		if (!businessCategoryDto.getName().equalsIgnoreCase(Constant.BUSINESS_CATEGORY_FOOD_DELIVERY)) {
 			productResponseDTO.setProductOutOfStock(availableQty > 0);
@@ -644,9 +643,9 @@ public class ProductServiceImpl implements ProductService {
 	}
 
 	/**
-	 * @param active
-	 * @param existingProduct
-	 * @param productId
+	 * @param  active
+	 * @param  existingProduct
+	 * @param  productId
 	 * @throws NotFoundException
 	 * @throws ValidationException
 	 */
@@ -676,7 +675,7 @@ public class ProductServiceImpl implements ProductService {
 	/**
 	 * check masters for activate product
 	 *
-	 * @param existingProduct
+	 * @param  existingProduct
 	 * @throws ValidationException
 	 * @throws NotFoundException
 	 */
@@ -706,8 +705,8 @@ public class ProductServiceImpl implements ProductService {
 
 		UserLogin userLogin = getUserLoginFromToken();
 		/**
-		 * In case of customer the userLogin might be anonymous user, resulting in no user login and hence if the userLogin is
-		 * null or the userLogin->entityType is customer then we will give records specific to customer
+		 * In case of customer the userLogin might be anonymous user, resulting in no user login and hence if the userLogin is null or the userLogin->entityType
+		 * is customer then we will give records specific to customer
 		 */
 		Boolean listForAdmin = true;
 		if (userLogin != null && UserType.VENDOR.name().equals(userLogin.getEntityType())) {
@@ -723,8 +722,8 @@ public class ProductServiceImpl implements ProductService {
 	/**
 	 * upload image of product
 	 *
-	 * @param image
-	 * @param product
+	 * @param  image
+	 * @param  product
 	 * @throws ValidationException
 	 * @throws FileOperationException
 	 */
@@ -844,8 +843,8 @@ public class ProductServiceImpl implements ProductService {
 	}
 
 	/**
-	 * @param productImportDTOs
-	 * @param userId
+	 * @param  productImportDTOs
+	 * @param  userId
 	 * @return
 	 */
 	private List<ProductImportDTO> insertListOfProducts(final List<ProductImportDTO> productImportDTOs) {
@@ -917,10 +916,10 @@ public class ProductServiceImpl implements ProductService {
 	}
 
 	/**
-	 * @param productImportDTO
-	 * @param brandId
-	 * @param vendor
-	 * @param productRequestDTO
+	 * @param  productImportDTO
+	 * @param  brandId
+	 * @param  vendor
+	 * @param  productRequestDTO
 	 * @throws ValidationException
 	 */
 	private void validateAndSetProductImport(final ProductImportDTO productImportDTO, final Long brandId, final Vendor vendor,
@@ -977,6 +976,9 @@ public class ProductServiceImpl implements ProductService {
 	@Override
 	public void deleteImage(final String imageType, final Long productId) throws NotFoundException, ValidationException {
 		Product product = getProductDetail(productId);
+		if (product.getDetailImage().equals(productDetailImage) || product.getImage().equals(productListImage)) {
+			throw new ValidationException(messageByLocaleService.getMessage("default.image.can.not.delete", null));
+		}
 		if (imageType.equals("DETAIL")) {
 			deleteOldImage(product.getDetailImage());
 			product.setDetailImage(null);
